Moving commit related stuff into feature folder
