+++ app/src/lib/commit/CommitCard.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
 	import CommitDragItem from './CommitDragItem.svelte';
-	import Icon from './Icon.svelte';
-	import BranchFilesList from '../file/BranchFilesList.svelte';
 	import { Project } from '$lib/backend/projects';
+	import CommitMessageInput from '$lib/commit/CommitMessageInput.svelte';
 	import Button from '$lib/components/Button.svelte';
-	import CommitMessageInput from '$lib/components/CommitMessageInput.svelte';
+	import Icon from '$lib/components/Icon.svelte';
 	import Modal from '$lib/components/Modal.svelte';
 	import { persistedCommitMessage } from '$lib/config/config';
 	import { draggable } from '$lib/dragging/draggable';
 	import { DraggableCommit, nonDraggable } from '$lib/dragging/draggables';
+	import BranchFilesList from '$lib/file/BranchFilesList.svelte';
 	import { copyToClipboard } from '$lib/utils/clipboard';
 	import { getContext, getContextStore } from '$lib/utils/context';
 	import { getTimeAgo } from '$lib/utils/timeAgo';
