+++ app/src/lib/components/RemoteBranchPreview.svelte
@@ -3,7 +3,7 @@
 	import Resizer from './Resizer.svelte';
 	import ScrollableContainer from './ScrollableContainer.svelte';
 	import { Project } from '$lib/backend/projects';
-	import CommitCard from '$lib/components/CommitCard.svelte';
+	import CommitCard from '$lib/commit/CommitCard.svelte';
 	import FileCard from '$lib/file/FileCard.svelte';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import { getRemoteBranchData } from '$lib/stores/remoteBranches';
